apps/discord_bot/pyproject.toml
[project]
name = "discord-bot-app"
version = "0.1.0"
description = "Discord gateway service for 508.dev"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.1",
    "discord.py~=2.6.0",
    "pydantic~=2.10",
    "pydantic-settings~=2.8",
    "redis>=6.4.0",
    "requests~=2.31",
    "rq>=2.6.0",
    "five08",
]

[tool.uv.sources]
five08 = { workspace = true }

[project.scripts]
discord-bot = "five08.discord_bot.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/five08"]
